chore(core): move error definitions & leave more comments
