+++ packages/nx/src/native/index.js
@@ -4,7 +4,7 @@ const Module = require('module');
 const { nxVersion} = require("../utils/versions")
 const { cacheDir} = require("../utils/cache-directory")
 
-const nxPackages = [
+const nxPackages = new Set([
   '@nx/nx-android-arm64',
   '@nx/nx-android-arm-eabi',
   '@nx/nx-win32-x64-msvc',
@@ -19,7 +19,7 @@ const nxPackages = [
   '@nx/nx-linux-arm64-musl',
   '@nx/nx-linux-arm64-gnu',
   '@nx/nx-linux-arm-gnueabihf',
-];
+]);
 
 const localNodeFiles = [
   'nx.android-arm64.node',
@@ -40,14 +40,19 @@ const localNodeFiles = [
 
 const originalLoad = Module._load;
 
+// We override the _load function so that when a native file is required, 
+// we copy it to a cache directory and require it from there.
+// This prevents the file being loaded from node_modules and causing file locking issues.
+// Will only be called once because the require cache takes over afterwards.
 Module._load = function (request, parent, isMain) {
   const modulePath = request;
   if (
-    nxPackages.includes(modulePath) ||
+    nxPackages.has(modulePath) ||
     localNodeFiles.some((f) => modulePath.endsWith(f))
   ) {
     const nativeLocation = require.resolve(modulePath);
     const fileName = basename(nativeLocation)
+    // we copy the file to the cache directory (.nx/cache by default) and prefix with nxVersion to avoid stale files being loaded
     const tmpFile = join(cacheDir, nxVersion + '-' + fileName);
     if (existsSync(tmpFile)) {
       return originalLoad.apply(this, [tmpFile, parent, isMain]);
@@ -58,6 +63,7 @@ Module._load = function (request, parent, isMain) {
     copyFileSync(nativeLocation, tmpFile);
     return originalLoad.apply(this, [tmpFile, parent, isMain]);
   } else {
+    // call the original _load function for everything else
     return originalLoad.apply(this, arguments);
   }
 };
